refactored plus/minus functionality to RetailerView, check button adds customer to store count
